+++ CHANGES.md
@@ -1,3 +1,9 @@
+2.0.15
+ * Option added to automatically create a database table
+
+2.0.13
+ * Ability to add additional properties as columns in the database
+
 2.0.1
  * Option to [write times in UTC](https://github.com/serilog/serilog-sinks-mssqlserver/pull/1)